requirements.txt
fastapi
uvicorn
twilio
python-multipart
python-dotenv
openai

